Rename UserServiceTest, fix test not working
