+++ src/test/kotlin/ar/edu/unq/criptop2p/Criptop2pApplicationTests.kt
@@ -2,8 +2,10 @@ package ar.edu.unq.criptop2p
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
 
 @SpringBootTest
+@ActiveProfiles(profiles = ["test"])
 class Criptop2pApplicationTests {
 
 	@Test
