+++ src/test/kotlin/ar/edu/unq/criptop2p/controller/UserControllerTest.kt
@@ -1,10 +1,10 @@
-package ar.edu.unq.criptop2p.service
+package ar.edu.unq.criptop2p.controller
 
-import ar.edu.unq.criptop2p.controller.UserController
 import ar.edu.unq.criptop2p.controller.dto.LoginDTO
 import ar.edu.unq.criptop2p.controller.dto.UserDTO
 import ar.edu.unq.criptop2p.model.Address
 import ar.edu.unq.criptop2p.persistance.UserRepository
+import ar.edu.unq.criptop2p.service.UserService
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.hamcrest.Matchers
 import org.junit.jupiter.api.Test
@@ -14,6 +14,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.http.MediaType
 import org.springframework.test.annotation.DirtiesContext
+import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath
@@ -21,6 +22,7 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 
 @AutoConfigureMockMvc
 @SpringBootTest
+@ActiveProfiles(profiles = ["test"])
 class userServiceTest (
         @Autowired
         private val mvc: MockMvc,
